Adapted changes from the 19/longer-cache-lines branch for master
This branch was initially started to increase the cache line size. However it
quickly became used for general improvements to the system. The new DRAM struct
design in this branch is being put on hold, however the general improvmenets
made in this branch are valuable, and should be merged into master.